+++ src/Plivo/Resources/PhoneNumber/PhoneNumberInterface.php
@@ -90,14 +90,14 @@ class PhoneNumberInterface extends ResourceInterface
      * @param number $phoneNumber
      * @param string|null $appId
      * @param string|null $cnamLookup
-     * @param bool|null $haEnabled
+     * @param bool|null $haEnable
      * @return PhoneNumberBuyResponse output
      */
-    public function buy($phoneNumber, $appId = null, $cnamLookup = null, $haEnabled = null)
+    public function buy($phoneNumber, $appId = null, $cnamLookup = null, $haEnable = null)
     {
         $data = ['app_id'=>$appId,'cnam_lookup'=>$cnamLookup];
-        if (!is_null($haEnabled)) {
-            $data['ha_enabled'] = $haEnabled;
+        if (!is_null($haEnable)) {
+            $data['ha_enable'] = $haEnable;
         }
         $response = $this->client->update(
             $this->uri . $phoneNumber . '/',
@@ -106,6 +106,10 @@ class PhoneNumberInterface extends ResourceInterface
 
         $responseContents = $response->getContent();
         if(!array_key_exists("error",$responseContents)){
+            $fallbackNumber = null;
+            if (isset($responseContents['numbers'][0]['fallback_number'])) {
+                $fallbackNumber = $responseContents['numbers'][0]['fallback_number'];
+            }
             return new PhoneNumberBuyResponse(
                 $responseContents['api_id'],
                 $responseContents['message'],
@@ -113,7 +117,7 @@ class PhoneNumberInterface extends ResourceInterface
                 $responseContents['numbers'][0]['status'],
                 $responseContents['status'],
                 $response->getStatusCode(),
-                $responseContents['numbers'][0]['fallback_number'] ?? null
+                $fallbackNumber
             );
         } elseif (gettype($responseContents['error']) == "array" && array_key_exists("error",$responseContents['error'])) {
             throw new PlivoResponseException(